% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policyfunctions.R
\name{Balanced}
\alias{Balanced}
\title{Picks an action using a balanced experiment approach.'}
\usage{
Balanced(t, TestPrices)
}
\arguments{
\item{t}{Current iteration of experiment}

\item{TestPrices}{Set of prices being tested}
}
\value{
An index, k, corresponding to the action chosen
}
\description{
Picks an action using a balanced experiment approach.'
}
