% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{CredibleInterval}
\alias{CredibleInterval}
\title{Creates 95 percent Credible Intervals for the probability of purchase given data.'}
\usage{
CredibleInterval(TestPrices, ArmDemands)
}
\arguments{
\item{TestPrices:}{Vector of prices being considered}

\item{ArmDemands:}{List of vectors containing complete purchase decisions for each price}
}
\value{
A list containing 2 vectors: low credible bounds and high credible bounds
}
\description{
Creates 95 percent Credible Intervals for the probability of purchase given data.'
}
