% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{LearningGraph}
\alias{LearningGraph}
\title{Plots Learning of demand or rewards by each policy'}
\usage{
LearningGraph(
  TestPrices,
  ArmDemands,
  Color,
  Title,
  DemandOrProfit,
  TrueResults,
  ylab
)
}
\arguments{
\item{TestPrices}{Set of prices being tested}

\item{ArmDemands}{List of vectors containing complete purchase decisions for each price}

\item{Color}{Color of the dots}

\item{Title}{Title of the graph}

\item{DemandOrProfit}{"Demand" or "Rewards" depending on what it is learning}

\item{TrueResults}{The True demand or rewards function}

\item{ylab}{y-axis label}
}
\value{
A performance plot
}
\description{
Plots Learning of demand or rewards by each policy'
}
