% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mab.R
\name{MAB}
\alias{MAB}
\title{Picks an action using a balanced experiment approach.'}
\usage{
MAB(Policy, TestPrices, NumIter, ShapeA, ShapeB, Seed)
}
\arguments{
\item{Policy}{Price decision rule (choose Balanced1, UCB1 or TS1)}

\item{TestPrices}{Set of prices being tested}

\item{NumIter}{Number of customers in the experiment}

\item{ShapeA}{Shape parameter of distribution}

\item{ShapeB}{Shape parameter of distribution}

\item{Seed}{Seed number for replication}
}
\value{
List containing ArmDemands, Actions, Rewards, Demands
}
\description{
Picks an action using a balanced experiment approach.'
}
