% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalchoice.R
\name{OptimalChoice}
\alias{OptimalChoice}
\title{Finds the optimal price and corresponding reward given a set of prices and
true unknown utility distribution (optimal is taken from set of prices).'}
\usage{
OptimalChoice(TestPrices, ShapeA, ShapeB)
}
\arguments{
\item{TestPrices}{Set of prices being tested}

\item{ShapeA}{Shape parameter of distribution}

\item{ShapeB}{Shape parameter of distribution}
}
\value{
Vector containing the optimal price and corresponding reward
}
\description{
Finds the optimal price and corresponding reward given a set of prices and
true unknown utility distribution (optimal is taken from set of prices).'
}
