% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PlotHistogram}
\alias{PlotHistogram}
\title{Plots Histogram of Prices Played'}
\usage{
PlotHistogram(
  Results1,
  Results2,
  Results3,
  NumIter,
  TestPrices,
  Optimal,
  Title
)
}
\arguments{
\item{Results1}{Balanced experiment results}

\item{Results2}{UCB experiment results}

\item{Results3}{TS experiment results}

\item{NumIter}{Number of customers in the experiment}

\item{TestPrices}{Set of prices being tested}

\item{Optimal}{The optimal price for the experiment}

\item{Title}{Title for the Graph}
}
\value{
A histogram plot
}
\description{
Plots Histogram of Prices Played'
}
