% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PlotPerformance}
\alias{PlotPerformance}
\title{Plots Performance of Various policies'}
\usage{
PlotPerformance(Rewards1, Rewards2, Rewards3, NumIter, Optimal, Title)
}
\arguments{
\item{NumIter}{Number of customers in the experiment}

\item{Optimal}{The optimal reward for the experiment}

\item{Title}{Title for the Graph}

\item{Results1}{Balanced experiment results}

\item{Results2}{UCB experiment results}

\item{Results3}{TS experiment results}
}
\value{
A performance plot
}
\description{
Plots Performance of Various policies'
}
