% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PosteriorLikelihoods}
\alias{PosteriorLikelihoods}
\title{Finds the likelihood of the true probability of purchase given experimental data
(recall purchasing is binary decision)}
\usage{
PosteriorLikelihoods(Successes, TimesTested)
}
\arguments{
\item{Successes}{Number of times that a purchase was made at the price tested}

\item{TimesTested}{Number of times a price has been tested}
}
\value{
Vector containing likelihoods of purchase probability over a fine grid
}
\description{
Finds the likelihood of the true probability of purchase given experimental data
(recall purchasing is binary decision)
}
