% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policyfunctions.R
\name{TS}
\alias{TS}
\title{Picks an action given past experiment data using Bernoulli Thompson Sampling.'}
\usage{
TS(ArmDemands, TestPrices)
}
\arguments{
\item{ArmDemands}{List of vectors containing complete purchase decisions for each price}

\item{TestPrices}{Set of prices being tested}
}
\value{
An index, k, corresponding to the action chosen
}
\description{
Picks an action given past experiment data using Bernoulli Thompson Sampling.'
}
