% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policyfunctions.R
\name{UCB}
\alias{UCB}
\title{Picks a price given past experiment data using the UCB1-Tuned-Price Algorithm.'}
\usage{
UCB(ArmDemands, TestPrices)
}
\arguments{
\item{ArmDemands}{List of vectors containing complete purchase decisions for each price}

\item{TestPrices}{Set of prices being tested}
}
\value{
An index, k, corresponding to the action chosen
}
\description{
Picks a price given past experiment data using the UCB1-Tuned-Price Algorithm.'
}
